+++ components/persistor/lib/api.ts
@@ -10,11 +10,28 @@
  *
  */
 
+
 module.exports = function (PersistObjectTemplate, baseClassForPersist) {
+    let supertypeRequire = require('@havenlife/supertype');
+    let statsDHelper = supertypeRequire.StatsdHelper;
 
     var Promise = require('bluebird');
     var _ = require('underscore');
 
+
+    function getTime() {
+        return process.hrtime();
+    }
+
+    function getStats(startTime, templateName: string, queryType: string, error = false) {
+        return statsDHelper.computeTimingAndSend(
+            startTime,
+            `persistor.${queryType}`,
+            {
+                error: error,
+                templateName: templateName,
+            });
+    }
 
     /**
      * PUBLIC INTERFACE FOR OBJECTS
@@ -116,15 +133,17 @@ module.exports = function (PersistObjectTemplate, baseClassForPersist) {
     }
     PersistObjectTemplate._injectTemplateFunctions = function (template) {
         function logExceptionAndRethrow(exception, logger, template, query, activity) {
-            if (typeof(query) === 'undefined') {
+            if (typeof (query) === 'undefined') {
                 query = 'undefined value provided';
-            } else if (typeof(query) === 'object') {
-                let undefHandler = (key, value) => typeof(value)  === 'undefined' ? 'undefined value provided for ' + key : value;
+            } else if (typeof (query) === 'object') {
+                let undefHandler = (key, value) => typeof (value) === 'undefined' ? 'undefined value provided for ' + key : value;
                 query = JSON.stringify(query, undefHandler);
             }
 
-            logger.error({component: 'persistor', module: 'api', activity: activity,
-                data: {template: template, query: query}});
+            logger.error({
+                component: 'persistor', module: 'api', activity: activity,
+                data: { template: template, query: query }
+            });
             throw exception;
         }
 
@@ -139,18 +158,31 @@ module.exports = function (PersistObjectTemplate, baseClassForPersist) {
          * @param {object} logger objecttemplate logger
          * @returns {object}
          */
-        template.getFromPersistWithId = async function(id, cascade, isTransient, idMap, isRefresh, logger) {
-            (logger || PersistObjectTemplate.logger).debug({component: 'persistor', module: 'api', activity: 'getFromPersistWithId',
-                data: {template: template.__name__, id: id}});
+        template.getFromPersistWithId = async function (id, cascade, isTransient, idMap, isRefresh, logger) {
+            (logger || PersistObjectTemplate.logger).debug({
+                component: 'persistor', module: 'api', activity: 'getFromPersistWithId',
+                data: { template: template.__name__, id: id }
+            });
             var dbType = PersistObjectTemplate.getDB(PersistObjectTemplate.getDBAlias(template.__collection__)).type;
+            const time = getTime();
             let getQuery = (dbType == PersistObjectTemplate.DB_Mongo ?
                 PersistObjectTemplate.getFromPersistWithMongoId(template, id, cascade, isTransient, idMap, logger) :
                 PersistObjectTemplate.getFromPersistWithKnexId(template, id, cascade, isTransient, idMap, isRefresh, logger))
-                .then(function(res) {
+                .then(function (res) {
                     return res;
                 }.bind(this));
 
-            return getQuery.catch(e => logExceptionAndRethrow(e, logger || PersistObjectTemplate.logger, template.__name__, id, 'getFromPersistWithId'));
+            const name = 'getFromPersistWithId';
+            return getQuery
+                .then(result => {
+                    getStats(time, template.__name__, name);
+                    return result;
+                })
+                .catch(e => {
+                    getStats(time, template.__name__, name, true);
+                    return logExceptionAndRethrow(e, logger || PersistObjectTemplate.logger, template.__name__, id, name)
+                });
+
         };
 
         /**
@@ -167,18 +199,32 @@ module.exports = function (PersistObjectTemplate, baseClassForPersist) {
          * @returns {object}
          * @deprecated in favor of persistorFetchWithQuery
          */
-        template.getFromPersistWithQuery = async function(query, cascade, start, limit, isTransient, idMap, options, logger) {
-            (logger || PersistObjectTemplate.logger).debug({component: 'persistor', module: 'api', activity: 'getFromPersistWithQuery',
-                data: {template: template.__name__}});
+        template.getFromPersistWithQuery = async function (query, cascade, start, limit, isTransient, idMap, options, logger) {
+            (logger || PersistObjectTemplate.logger).debug({
+                component: 'persistor', module: 'api', activity: 'getFromPersistWithQuery',
+                data: { template: template.__name__ }
+            });
             var dbType = PersistObjectTemplate.getDB(PersistObjectTemplate.getDBAlias(template.__collection__)).type;
+            const time = getTime();
+
             let getQuery = (dbType == PersistObjectTemplate.DB_Mongo ?
                 PersistObjectTemplate.getFromPersistWithMongoQuery(template, query, cascade, start, limit, isTransient, idMap, options, logger) :
                 PersistObjectTemplate.getFromPersistWithKnexQuery(null, template, query, cascade, start, limit, isTransient, idMap, options, undefined, undefined, logger))
-                .then(function(res) {
+                .then(function (res) {
                     return res;
                 }.bind(this));
 
-            return getQuery.catch(e => logExceptionAndRethrow(e, logger || PersistObjectTemplate.logger, template.__name__, query, 'getFromPersistWithQuery'));
+
+            const name = 'getFromPersistWithQuery';
+            return getQuery
+                .then(result => {
+                    getStats(time, template.__name__, name);
+                    return result;
+                })
+                .catch(e => {
+                    getStats(time, template.__name__, name, true);
+                    return logExceptionAndRethrow(e, logger || PersistObjectTemplate.logger, template.__name__, query, name)
+                });
         };
 
         /**
@@ -190,11 +236,25 @@ module.exports = function (PersistObjectTemplate, baseClassForPersist) {
          * @returns {object}
          * @deprecated in favor of persitorDeleteByQuery
          */
-        template.deleteFromPersistWithQuery = async function(query, txn, logger) {
+        template.deleteFromPersistWithQuery = async function (query, txn, logger) {
             var dbType = PersistObjectTemplate.getDB(PersistObjectTemplate.getDBAlias(template.__collection__)).type;
-            return dbType == PersistObjectTemplate.DB_Mongo ?
+            const time = getTime();
+
+            let getQuery = dbType == PersistObjectTemplate.DB_Mongo ?
                 PersistObjectTemplate.deleteFromPersistWithMongoQuery(template, query, logger) :
                 PersistObjectTemplate.deleteFromKnexQuery(template, query, txn, logger);
+
+            const name = 'deleteFromQuery';
+            return getQuery
+                .then(result => {
+                    getStats(time, template.__name__, name);
+                    return result;
+                })
+                // @TODO: need to handle errors with log
+
+                .catch(e => {
+                    getStats(time, template.__name__, name, true);
+                });
         };
 
         /**
@@ -203,7 +263,9 @@ module.exports = function (PersistObjectTemplate, baseClassForPersist) {
          * @param {json} options @todo
          * @returns {*}
          */
-        template.persistorFetchById = async function(id, options) { // @TODO: Legacy
+        template.persistorFetchById = async function (id, options) { // @TODO: Legacy
+            const time = getTime();
+
             PersistObjectTemplate._validateParams(options, 'fetchSchema', template);
 
             options = options || {};
@@ -211,13 +273,23 @@ module.exports = function (PersistObjectTemplate, baseClassForPersist) {
             var persistObjectTemplate = options.session || PersistObjectTemplate;
             (options.logger || persistObjectTemplate.logger).debug({
                 component: 'persistor', module: 'api', activity: 'persistorFetchById',
-                data: {template: template.__name__, id: id}
+                data: { template: template.__name__, id: id }
             });
             var dbType = persistObjectTemplate.getDB(persistObjectTemplate.getDBAlias(template.__collection__)).type;
             let fetchQuery = (dbType == persistObjectTemplate.DB_Mongo ?
                 persistObjectTemplate.getFromPersistWithMongoId(template, id, options.fetch, options.transient, null, options.logger) :
                 persistObjectTemplate.getFromPersistWithKnexId(template, id, options.fetch, options.transient, null, null, options.logger, options.enableChangeTracking, options.projection));
-            return fetchQuery.catch(e => logExceptionAndRethrow(e, options.logger || persistObjectTemplate.logger, template.__name__, id, 'persistorFetchById'));
+
+            const name = 'persistorFetchById';
+            return fetchQuery
+                .then(result => {
+                    getStats(time, template.__name__, name);
+                    return result;
+                })
+                .catch(e => {
+                    getStats(time, template.__name__, name, true);
+                    return logExceptionAndRethrow(e, options.logger || PersistObjectTemplate.logger, template.__name__, id, name)
+                });
         };
 
         /**
@@ -226,14 +298,28 @@ module.exports = function (PersistObjectTemplate, baseClassForPersist) {
          * @param {JSON} options @TODO
          * @returns {Object}
          */
-        template.persistorDeleteByQuery = async function(query, options) {
+        template.persistorDeleteByQuery = async function (query, options) {
+            const time = getTime();
+
             PersistObjectTemplate._validateParams(options, 'persistSchema', template);
 
             options = options || {};
             var dbType = PersistObjectTemplate.getDB(PersistObjectTemplate.getDBAlias(template.__collection__)).type;
-            return dbType == PersistObjectTemplate.DB_Mongo ?
-                        PersistObjectTemplate.deleteFromPersistWithMongoQuery(template, query, options.logger) :
-                        PersistObjectTemplate.deleteFromKnexByQuery(template, query, options.transaction, options.logger);
+            let deleteQuery = dbType == PersistObjectTemplate.DB_Mongo ?
+                PersistObjectTemplate.deleteFromPersistWithMongoQuery(template, query, options.logger) :
+                PersistObjectTemplate.deleteFromKnexByQuery(template, query, options.transaction, options.logger);
+
+            const name = 'persistorDeleteByQuery';
+            return deleteQuery
+                .then(result => {
+                    getStats(time, template.__name__, name);
+                    return result;
+                })
+                // @TODO: need to handle errors with log
+                .catch(e => {
+                    getStats(time, template.__name__, name, true);
+                    return logExceptionAndRethrow(e, options.logger || PersistObjectTemplate.logger, template.__name__, query, name);
+                });
         };
 
         /**
@@ -242,25 +328,39 @@ module.exports = function (PersistObjectTemplate, baseClassForPersist) {
          * @param {JSON} options @TODO
          * @returns {*}
          */
-        template.persistorFetchByQuery = async function(query, options) {
+        template.persistorFetchByQuery = async function (query, options) {
+            const time = getTime();
+
             PersistObjectTemplate._validateParams(options, 'fetchSchema', template);
 
             options = options || {};
             var persistObjectTemplate = options.session || PersistObjectTemplate;
             var logger = options.logger || persistObjectTemplate.logger;
-            logger.debug({component: 'persistor', module: 'api', activity: 'getFromPersistWithQuery',
-                data: {template: template.__name__}});
+            logger.debug({
+                component: 'persistor', module: 'api', activity: 'persistorFetchByQuery',
+                data: { template: template.__name__ }
+            });
             var dbType = persistObjectTemplate.getDB(persistObjectTemplate.getDBAlias(template.__collection__)).type;
             if (options.order && !options.order.sort) {
                 options.order = { sort: options.order };
             }
             let fetchQuery = (dbType == persistObjectTemplate.DB_Mongo ?
                 persistObjectTemplate.getFromPersistWithMongoQuery(template, query, options.fetch, options.start,
-                            options.limit, options.transient, options.order, options.order, logger) :
+                    options.limit, options.transient, options.order, options.order, logger) :
                 persistObjectTemplate.getFromPersistWithKnexQuery(null, template, query, options.fetch, options.start,
-                            options.limit, options.transient, null, options.order,
-                            undefined, undefined, logger, options.enableChangeTracking, options.projection));
-            return fetchQuery.catch(e => logExceptionAndRethrow(e, options.logger || PersistObjectTemplate.logger, template.__name__, query, 'persistorFetchByQuery'));
+                    options.limit, options.transient, null, options.order,
+                    undefined, undefined, logger, options.enableChangeTracking, options.projection));
+
+            const name = 'persistorFetchByQuery';
+            return fetchQuery
+                .then(result => {
+                    getStats(time, template.__name__, name);
+                    return result;
+                })
+                .catch(e => {
+                    getStats(time, template.__name__, name, true);
+                    return logExceptionAndRethrow(e, options.logger || PersistObjectTemplate.logger, template.__name__, query, name)
+                });
         };
         /**
          * Return count of objects of this class given a json query
@@ -269,22 +369,36 @@ module.exports = function (PersistObjectTemplate, baseClassForPersist) {
          * @param {object} options @TODO
          * @returns {Number}
          */
-        template.persistorCountByQuery = async function(query, options) {
+        template.persistorCountByQuery = async function (query, options) {
+            const time = getTime();
+
             PersistObjectTemplate._validateParams(options, 'fetchSchema', template);
 
             options = options || {};
             var logger = options.logger || PersistObjectTemplate.logger;
-            logger.debug({component: 'persistor', module: 'api', activity: 'getFromPersistWithQuery',
-                data: {template: template.__name__}});
+            logger.debug({
+                component: 'persistor', module: 'api', activity: 'persistorCountByQuery',
+                data: { template: template.__name__ }
+            });
 
             var dbType = PersistObjectTemplate.getDB(PersistObjectTemplate.getDBAlias(template.__collection__)).type;
             let countQuery = (dbType == PersistObjectTemplate.DB_Mongo ?
                 PersistObjectTemplate.countFromMongoQuery(template, query, logger) :
                 PersistObjectTemplate.countFromKnexQuery(template, query, logger))
-                .then(function(res) {
+                .then(function (res) {
                     return res;
                 }.bind(this));
-            return countQuery.catch(e => logExceptionAndRethrow(e, options.logger || PersistObjectTemplate.logger, template.__name__, query, {activity: 'persistorCountByQuery'}));
+
+            const name = 'persistorCountByQuery';
+            return countQuery
+                .then(result => {
+                    getStats(time, template.__name__, name);
+                    return result;
+                })
+                .catch(e => {
+                    getStats(time, template.__name__, name, true);
+                    return logExceptionAndRethrow(e, options.logger || PersistObjectTemplate.logger, template.__name__, query, { activity: 'persistorCountByQuery' })
+                });
         };
 
         /**
@@ -296,17 +410,31 @@ module.exports = function (PersistObjectTemplate, baseClassForPersist) {
          * @returns {object}
          * @deprecated in favor of persistorDeleteByQuery
          */
-        template.deleteFromPersistWithId = async function(id, txn, logger) {
-            (logger || PersistObjectTemplate.logger).debug({component: 'persistor', module: 'api', activity: 'deleteFromPersistWithId',
-                data: {template: template.__name__, id: id}});
+        template.deleteFromPersistWithId = async function (id, txn, logger) {
+            const time = getTime();
+
+            (logger || PersistObjectTemplate.logger).debug({
+                component: 'persistor', module: 'api', activity: 'deleteFromPersistWithId',
+                data: { template: template.__name__, id: id }
+            });
             var dbType = PersistObjectTemplate.getDB(PersistObjectTemplate.getDBAlias(template.__collection__)).type;
             let deleteQuery = (dbType == PersistObjectTemplate.DB_Mongo ?
                 PersistObjectTemplate.deleteFromPersistWithMongoId(template, id, logger) :
                 PersistObjectTemplate.deleteFromKnexId(template, id, txn, logger))
-                .then(function(res) {
+                .then(function (res) {
                     return res;
                 }.bind(this));
-            return deleteQuery.catch(e => logExceptionAndRethrow(e, logger || PersistObjectTemplate.logger, template.__name__, id, {activity: 'deleteFromPersistWithId'}));
+
+            const name = 'deleteFromPersistWithId';
+            return deleteQuery
+                .then(result => {
+                    getStats(time, template.__name__, name);
+                    return result;
+                })
+                .catch(e => {
+                    getStats(time, template.__name__, name, true);
+                    return logExceptionAndRethrow(e, logger || PersistObjectTemplate.logger, template.__name__, id, { activity: 'deleteFromPersistWithId' })
+                });
         };
 
         /**
@@ -317,17 +445,31 @@ module.exports = function (PersistObjectTemplate, baseClassForPersist) {
          * @returns {Number}
          * @deprecated in favor of persistorCountWithQuery
          */
-        template.countFromPersistWithQuery = async function(query, logger) {
-            (logger || PersistObjectTemplate.logger).debug({component: 'persistor', module: 'api', activity: 'countFromPersistWithQuery',
-                data: {template: template.__name__}});
+        template.countFromPersistWithQuery = async function (query, logger) {
+            const time = getTime();
+
+            (logger || PersistObjectTemplate.logger).debug({
+                component: 'persistor', module: 'api', activity: 'countFromPersistWithQuery',
+                data: { template: template.__name__ }
+            });
             var dbType = PersistObjectTemplate.getDB(PersistObjectTemplate.getDBAlias(template.__collection__)).type;
             let countQuery = (dbType == PersistObjectTemplate.DB_Mongo ?
                 PersistObjectTemplate.countFromMongoQuery(template, query, logger) :
                 PersistObjectTemplate.countFromKnexQuery(template, query, logger))
-                .then(function(res) {
+                .then(function (res) {
                     return res;
                 }.bind(this));
-            return countQuery.catch(e => logExceptionAndRethrow(e, logger || PersistObjectTemplate.logger, template.__name__, query, 'countFromPersistWithQuery'));
+
+            const name = 'countFromPersistWithQuery';
+            return countQuery
+                .then(result => {
+                    getStats(time, template.__name__, name);
+                    return result;
+                })
+                .catch(e => {
+                    getStats(time, template.__name__, name, true);
+                    return logExceptionAndRethrow(e, logger || PersistObjectTemplate.logger, template.__name__, query, name)
+                });
         };
 
         /**
@@ -364,7 +506,7 @@ module.exports = function (PersistObjectTemplate, baseClassForPersist) {
          * @returns {string}
          */
         template.persistorGetParentKey = function (prop, alias) {
-            return (alias ? alias + '.'  : '') + template.__schema__.parents[prop].id;
+            return (alias ? alias + '.' : '') + template.__schema__.parents[prop].id;
         };
 
         /**
@@ -374,7 +516,7 @@ module.exports = function (PersistObjectTemplate, baseClassForPersist) {
          * @returns {string}
          */
         template.persistorGetChildKey = function (prop, alias) {
-            return (alias ? alias + '.'  : '') + template.__schema__.children[prop].id;
+            return (alias ? alias + '.' : '') + template.__schema__.children[prop].id;
         };
 
         /**
@@ -383,7 +525,7 @@ module.exports = function (PersistObjectTemplate, baseClassForPersist) {
          * @returns {string}
          */
         template.persistorGetId = function (alias) {
-            return (alias ? alias + '.'  : '') + '_id';
+            return (alias ? alias + '.' : '') + '_id';
         };
 
         /**
@@ -445,20 +587,24 @@ module.exports = function (PersistObjectTemplate, baseClassForPersist) {
                             var closureQueryOptions = defineProperty.queryOptions ? defineProperty.queryOptions : {};
                             var toClient = !(defineProperty.isLocal || (defineProperty.toClient === false))
                             if (!props[closureProp + 'Persistor']) {
-                                template.createProperty(closureProp + 'Persistor', {type: Object, toClient: toClient,
+                                template.createProperty(closureProp + 'Persistor', {
+                                    type: Object, toClient: toClient,
                                     toServer: false, persist: false,
-                                    value: {isFetched: defineProperty.autoFetch ? false : true, isFetching: false}});
+                                    value: { isFetched: defineProperty.autoFetch ? false : true, isFetching: false }
+                                });
                             }
                             if (!template.prototype[closureProp + 'Fetch'])
-                                template.createProperty(closureProp + 'Fetch', {on: 'server', body: function (start, limit) {
-                                    if (typeof(start) != 'undefined') {
-                                        closureQueryOptions['skip'] = start;
+                                template.createProperty(closureProp + 'Fetch', {
+                                    on: 'server', body: function (start, limit) {
+                                        if (typeof (start) != 'undefined') {
+                                            closureQueryOptions['skip'] = start;
+                                        }
+                                        if (typeof (limit) != 'undefined') {
+                                            closureQueryOptions['limit'] = limit;
+                                        }
+                                        return this.fetchProperty(closureProp, closureFetch, closureQueryOptions);
                                     }
-                                    if (typeof(limit) != 'undefined') {
-                                        closureQueryOptions['limit'] = limit;
-                                    }
-                                    return this.fetchProperty(closureProp, closureFetch, closureQueryOptions);
-                                }});
+                                });
                         })();
                     }
                 }
@@ -471,91 +617,168 @@ module.exports = function (PersistObjectTemplate, baseClassForPersist) {
         var self = this; // this is the objectTemplate for non-TS apps or daemons or other non-sessionized use-cases
 
         template.prototype.persistSave = // Legacy
-        function (txn, logger) {
-            var persistObjectTemplate = this.__objectTemplate__ || self;
-            (logger || persistObjectTemplate.logger).debug({component: 'persistor', module: 'api', activity: 'persistSave',
-                data: {template: this.__template__.__name__, id: this.__id__}});
-            var dbType = persistObjectTemplate.getDB(persistObjectTemplate.getDBAlias(this.__template__.__collection__)).type;
-            return dbType == persistObjectTemplate.DB_Mongo ?
-                persistObjectTemplate.persistSaveMongo(this, undefined, undefined, undefined, txn, logger)
-                    .then (function (obj) {
-                        if (txn) {
-                            persistObjectTemplate.saved(obj, txn);
-                        }
-                        return Promise.resolve(obj._id.toString())
+            function (txn, logger) {
+                const time = getTime();
+
+                var persistObjectTemplate = this.__objectTemplate__ || self;
+                (logger || persistObjectTemplate.logger).debug({
+                    component: 'persistor', module: 'api', activity: 'persistSave',
+                    data: { template: this.__template__.__name__, id: this.__id__ }
+                });
+                var dbType = persistObjectTemplate.getDB(persistObjectTemplate.getDBAlias(this.__template__.__collection__)).type;
+                let query = dbType == persistObjectTemplate.DB_Mongo ?
+                    persistObjectTemplate.persistSaveMongo(this, undefined, undefined, undefined, txn, logger)
+                        .then(function (obj) {
+                            if (txn) {
+                                persistObjectTemplate.saved(obj, txn);
+                            }
+                            return Promise.resolve(obj._id.toString())
+                        })
+                    : persistObjectTemplate.persistSaveKnex(this, txn, logger)
+                        .then(function (obj) {
+                            if (txn) {
+                                persistObjectTemplate.saved(obj, txn);
+                            }
+                            return Promise.resolve(obj._id.toString());
+                        });
+                const name = 'persistSave';
+                return query
+                    .then(result => {
+                        getStats(time, template.__name__, name);
+                        return result;
                     })
-                : persistObjectTemplate.persistSaveKnex(this, txn, logger)
-                    .then (function (obj) {
-                        if (txn) {
-                            persistObjectTemplate.saved(obj, txn);
-                        }
-                        return Promise.resolve(obj._id.toString());
+                    // @TODO: need to handle errors with log
+                    .catch(e => {
+                        getStats(time, template.__name__, name, true);
+                        throw e;
                     });
-        };
+            };
 
         template.prototype.persistTouch = // Legacy -- just use persistorSave
-        async function (txn, logger) {
-            var persistObjectTemplate = this.__objectTemplate__ || self;
-            (logger || persistObjectTemplate.logger).debug({component: 'persistor', module: 'api', activity: 'persistTouch',
-                data: {template: this.__template__.__name__, id: this.__id__}});
-            var dbType = persistObjectTemplate.getDB(persistObjectTemplate.getDBAlias(this.__template__.__collection__)).type;
-            return dbType == persistObjectTemplate.DB_Mongo ?
-                persistObjectTemplate.persistSaveMongo(this, undefined, undefined, undefined, txn, logger)
-                : persistObjectTemplate.persistTouchKnex(this, txn, logger);
-        };
+            async function (txn, logger) {
+                const time = getTime();
+                var persistObjectTemplate = this.__objectTemplate__ || self;
+                (logger || persistObjectTemplate.logger).debug({
+                    component: 'persistor', module: 'api', activity: 'persistTouch',
+                    data: { template: this.__template__.__name__, id: this.__id__ }
+                });
+                var dbType = persistObjectTemplate.getDB(persistObjectTemplate.getDBAlias(this.__template__.__collection__)).type;
+                let query = dbType == persistObjectTemplate.DB_Mongo ?
+                    persistObjectTemplate.persistSaveMongo(this, undefined, undefined, undefined, txn, logger)
+                    : persistObjectTemplate.persistTouchKnex(this, txn, logger);
+
+                const name = 'persistTouch';
+                return query
+                    .then(result => {
+                        getStats(time, template.__name__, name);
+                        return result;
+                    })
+                    // @TODO: need to handle errors with log
+                    .catch(e => {
+                        getStats(time, template.__name__, name, true);
+                        throw e;
+                    });
+            };
 
         //persistDelete is modified to support both legacy and V2, options this is passed for V2 as the first parameter.
         template.prototype.persistDelete = // Legacy
-        async function (txn, logger) {
-            var persistObjectTemplate = this.__objectTemplate__ || self;
-            if (!txn || (txn && txn.knex && txn.knex.transacting)) {
-                (logger || persistObjectTemplate.logger).debug({component: 'persistor', module: 'api', activity: 'persistDelete',
-                    data: {template: this.__template__.__name__, id: this.__id__}});
-                if (txn) {
-                    delete txn.dirtyObjects[this.__id__];
-                }
-                return this.__template__.deleteFromPersistWithId(this._id, txn, logger)
-            }
-            else {
-                //for V2 options are passed as the first paramter.
-                this.deleteV2.call(this, txn);
-            }
+            async function (txn, logger) {
+                const time = getTime();
 
-        };
+                var persistObjectTemplate = this.__objectTemplate__ || self;
+                let query;
+                if (!txn || (txn && txn.knex && txn.knex.transacting)) {
+                    (logger || persistObjectTemplate.logger).debug({
+                        component: 'persistor', module: 'api', activity: 'persistDelete',
+                        data: { template: this.__template__.__name__, id: this.__id__ }
+                    });
+                    if (txn) {
+                        delete txn.dirtyObjects[this.__id__];
+                    }
+                    query = this.__template__.deleteFromPersistWithId(this._id, txn, logger)
+                }
+                else {
+                    //for V2 options are passed as the first paramter.
+                    query = this.deleteV2.call(this, txn);
+                }
+
+                const name = 'persistDelete';
+                return query
+                    .then(result => {
+                        getStats(time, template.__name__, name);
+                        return result;
+                    })
+                    // @TODO: need to handle errors with log
+                    .catch(e => {
+                        getStats(time, template.__name__, name, true);
+                        throw e;
+                    });
+
+            };
 
         // Legacy
-        template.prototype.setDirty =  function (txn, onlyIfChanged, cascade, logger) {
+        template.prototype.setDirty = function (txn, onlyIfChanged, cascade, logger) {
             var persistObjectTemplate = this.__objectTemplate__ || self;
             persistObjectTemplate.setDirty(this, txn, onlyIfChanged, !cascade, logger);
         };
 
-        template.prototype.setAsDeleted =  function (txn, onlyIfChanged) {
+        template.prototype.setAsDeleted = function (txn, onlyIfChanged) {
             var persistObjectTemplate = this.__objectTemplate__ || self;
             persistObjectTemplate.setAsDeleted(this, txn, onlyIfChanged)
         };
 
         // Legacy
         template.prototype.cascadeSave = function (txn, logger) {
+            var time = getTime();
+
             var persistObjectTemplate = this.__objectTemplate__ || self;
-            persistObjectTemplate.setDirty(this, txn || persistObjectTemplate.currentTransaction, true, false, logger);
+            var query = Promise.resolve(persistObjectTemplate.setDirty(this, txn || persistObjectTemplate.currentTransaction, true, false, logger));
+
+            const name = 'cascadeSave';
+            return query
+                .then(result => {
+                    getStats(time, template.__name__, name);
+                    return result;
+                })
+                // @TODO: need to handle errors with log
+                .catch(e => {
+                    getStats(time, template.__name__, name, true);
+                    throw e;
+                });
         };
 
         template.prototype.isStale = // Legacy
-        template.prototype.persistorIsState = function () {
-            var persistObjectTemplate = this.__objectTemplate__ || self;
-            var dbType = persistObjectTemplate.getDB(persistObjectTemplate.getDBAlias(this.__template__.__collection__)).type;
-            return this.__template__.countFromPersistWithQuery(
-                {_id: (dbType == persistObjectTemplate.DB_Mongo) ? persistObjectTemplate.ObjectID(this._id.toString()) : this._id,
-                    __version__: this.__version__}).then(function(count) {
+            template.prototype.persistorIsState = function () {
+                var time = getTime();
+                var name = 'isStale';
+                var persistObjectTemplate = this.__objectTemplate__ || self;
+                var dbType = persistObjectTemplate.getDB(persistObjectTemplate.getDBAlias(this.__template__.__collection__)).type;
+                return this.__template__.countFromPersistWithQuery(
+                    {
+                        _id: (dbType == persistObjectTemplate.DB_Mongo) ? persistObjectTemplate.ObjectID(this._id.toString()) : this._id,
+                        __version__: this.__version__
+                    }).then(function (count) {
                         return !count
-                    }.bind(this));
-        };
+                    }.bind(this))
+                    .then(result => {
+                        getStats(time, template.__name__, name);
+                        return result;
+                    })
+                    // @TODO: need to handle errors with log
+                    .catch(e => {
+                        getStats(time, template.__name__, name, true);
+                        throw e;
+                    });
+            };
 
         // Legacy
         template.prototype.fetchProperty = async function (prop, cascade, queryOptions, isTransient, idMap, logger) {
             var persistObjectTemplate = this.__objectTemplate__ || self;
-            (logger || persistObjectTemplate.logger).debug({component: 'persistor', module: 'api', activity: 'fetchProperty',
-                data: {template: this.__template__.__name__, id: this.__id__}});
+            var time = getTime();
+            (logger || persistObjectTemplate.logger).debug({
+                component: 'persistor', module: 'api', activity: 'fetchProperty',
+                data: { template: this.__template__.__name__, id: this.__id__ }
+            });
             idMap = idMap || {};
             var properties = {};
             var objectProperties = this.__template__.getProperties();
@@ -567,16 +790,31 @@ module.exports = function (PersistObjectTemplate, baseClassForPersist) {
             cascadeTop[prop] = cascade || true;
 
             var dbType = persistObjectTemplate.getDB(persistObjectTemplate.getDBAlias(this.__template__.__collection__)).type;
-            return dbType == persistObjectTemplate.DB_Mongo ?
+            var promise;
+            promise = dbType == persistObjectTemplate.DB_Mongo ?
                 persistObjectTemplate.getTemplateFromMongoPOJO(this, this.__template__, null, null, idMap, cascadeTop, this, properties, isTransient, logger) :
                 persistObjectTemplate.getTemplateFromKnexPOJO(this, this.__template__, null, idMap, cascadeTop, isTransient, null, this, properties, undefined, undefined, undefined, logger);
+
+            var name = 'fetchProperty';
+            return promise.then(result => {
+                getStats(time, template.__name__, name);
+                return result;
+            })
+                // @TODO: need to handle errors with log
+                .catch(e => {
+                    getStats(time, template.__name__, name, true);
+                    throw e;
+                });
 
         };
 
         template.prototype.fetch = async function (cascade, isTransient, idMap, logger) {
             var persistObjectTemplate = this.__objectTemplate__ || self;
-            (logger || persistObjectTemplate.logger).debug({component: 'persistor', module: 'api', activity: 'fetch',
-                data: {template: this.__template__.__name__, id: this.__id__}});
+            var time = getTime();
+            (logger || persistObjectTemplate.logger).debug({
+                component: 'persistor', module: 'api', activity: 'fetch',
+                data: { template: this.__template__.__name__, id: this.__id__ }
+            });
             idMap = idMap || {};
 
             var properties = {}
@@ -586,24 +824,34 @@ module.exports = function (PersistObjectTemplate, baseClassForPersist) {
             }
             var dbType = persistObjectTemplate.getDB(persistObjectTemplate.getDBAlias(this.__template__.__collection__)).type;
             var previousDirtyTracking = persistObjectTemplate.__changeTracking__;
-            return (dbType == persistObjectTemplate.DB_Mongo ?
+            persistObjectTemplate.__changeTracking__ = false;
+            var promise = (dbType == persistObjectTemplate.DB_Mongo) ?
                 persistObjectTemplate.getTemplateFromMongoPOJO(this, this.__template__, null, null, idMap, cascade, this, properties, isTransient, logger) :
-                persistObjectTemplate.getTemplateFromKnexPOJO(this, this.__template__, null, idMap, cascade, isTransient, null, this, properties, undefined, undefined, undefined, logger))
-                .then(function (res) {
-                    return res;
+                persistObjectTemplate.getTemplateFromKnexPOJO(this, this.__template__, null, idMap, cascade, isTransient, null, this, properties, undefined, undefined, undefined, logger);
+            var name = 'fetch';
+            return promise.then(result => {
+                getStats(time, template.__name__, name);
+                return result;
+            })
+                // @TODO: need to handle errors with log
+                .catch(e => {
+                    getStats(time, template.__name__, name, true);
+                    throw e;
                 });
         };
 
-        template.prototype.persistorFetchReferences = template.prototype.fetchReferences = async function(options) {
+        template.prototype.persistorFetchReferences = template.prototype.fetchReferences = async function (options) {
             var persistObjectTemplate = this.__objectTemplate__ || self;
             persistObjectTemplate._validateParams(options, 'fetchSchema', this.__template__);
+            var time = getTime();
 
             options = options || {};
             var logger = options.logger || persistObjectTemplate.logger;
             logger.debug({
                 component: 'persistor', module: 'api', activity: 'fetchReferences',
-                data: {template: this.__template__.__name__, id: this.__id__}
+                data: { template: this.__template__.__name__, id: this.__id__ }
             });
+
 
             var properties = {}
             var objectProperties = this.__template__.getProperties();
@@ -611,15 +859,28 @@ module.exports = function (PersistObjectTemplate, baseClassForPersist) {
                 properties[prop] = objectProperties[prop];
             }
             var dbType = persistObjectTemplate.getDB(persistObjectTemplate.getDBAlias(this.__template__.__collection__)).type;
-            return (dbType == persistObjectTemplate.DB_Mongo ?
+            var promise = (dbType == persistObjectTemplate.DB_Mongo ?
                 persistObjectTemplate.getTemplateFromMongoPOJO(this, this.__template__, null, null, {},
                     options.fetch, this, properties, options.transient, logger) :
                 persistObjectTemplate.getTemplateFromKnexPOJO(this, this.__template__, null, {},
                     options.fetch, options.transient, null, this,
                     properties, undefined, undefined, undefined, logger));
+
+            var name = 'persistorFetchReferences';
+            return promise
+                .then(result => {
+                    getStats(time, template.__name__, name);
+                    return result;
+                })// @TODO: need to handle errors with log
+                .catch(e => {
+                    getStats(time, template.__name__, name, true);
+                    throw e;
+                });
         };
 
         template.prototype.persistorRefresh = template.prototype.refresh = async function (logger) {
+
+            // @TODO: need to add stats collection to this method for success and failure
             var persistObjectTemplate = this.__objectTemplate__ || self;
             (logger || persistObjectTemplate.logger).debug({component: 'persistor', module: 'api', activity: 'refresh',
                 data: {template: this.__template__.__name__, id: this.__id__}});
@@ -631,7 +892,9 @@ module.exports = function (PersistObjectTemplate, baseClassForPersist) {
 
         };
 
-        template.prototype.persistorSave =  template.prototype.persist = async function (options) {
+        template.prototype.persistorSave = template.prototype.persist = async function (options) {
+            var time = getTime();
+
             var persistObjectTemplate = this.__objectTemplate__ || self;
             persistObjectTemplate._validateParams(options, 'persistSchema', this.__template__);
 
@@ -641,28 +904,43 @@ module.exports = function (PersistObjectTemplate, baseClassForPersist) {
             var logger = options.logger || persistObjectTemplate.logger;
             logger.debug({
                 component: 'persistor', module: 'api', activity: 'save',
-                data: {template: this.__template__.__name__, id: this.__id__}
+                data: { template: this.__template__.__name__, id: this.__id__ }
             });
 
+            var promise;
             if (!txn) {
-                return this.persistSave(null, logger);
+                promise = this.persistSave(null, logger);
             }
             else {
-                return Promise.resolve(this.setDirty(txn, false, cascade, logger));
+                promise = Promise.resolve(this.setDirty(txn, false, cascade, logger));
             }
+
+            const name = 'persistorSave';
+            return promise
+                .then(result => {
+                    getStats(time, template.__name__, name);
+                    return result;
+                })
+                // @TODO: need to handle errors with log
+
+                .catch(e => {
+                    getStats(time, template.__name__, name, true);
+                    throw e;
+                });
         };
 
         /**
          * Can generate object id even before saving the record to the database.
          * @returns {string}
          */
-        template.prototype.generateId =  function (): string {
+        template.prototype.generateId = function (): string {
             let persistObjectTemplate = this.__objectTemplate__ || self;
             return (this._id = this._id || persistObjectTemplate.createPrimaryKey(this));
         };
 
         //persistorDelete will only support new API calls.
-        template.prototype.persistorDelete = template.prototype.deleteV2 = async function(options) {
+        template.prototype.persistorDelete = template.prototype.deleteV2 = async function (options) {
+            var time = getTime();
             var persistObjectTemplate = this.__objectTemplate__ || self;
             persistObjectTemplate._validateParams(options, 'persistSchema', this.__template__);
 
@@ -672,19 +950,32 @@ module.exports = function (PersistObjectTemplate, baseClassForPersist) {
             var logger = options.logger || persistObjectTemplate.logger;
             logger.debug({
                 component: 'persistor', module: 'api', activity: 'delete',
-                data: {template: this.__template__.__name__, id: this.__id__}
+                data: { template: this.__template__.__name__, id: this.__id__ }
             });
 
+            var promise;
             if (!txn) {
-                return this.__template__.deleteFromPersistWithId(this._id, null, logger)
+                promise = this.__template__.deleteFromPersistWithId(this._id, null, logger)
             }
             else {
-                return Promise.resolve(persistObjectTemplate.setAsDeleted(this, txn, false, !cascade, logger));
+                promise = Promise.resolve(persistObjectTemplate.setAsDeleted(this, txn, false, !cascade, logger));
             }
+
+            const name = 'persistorDelete';
+            return promise.then(result => {
+                getStats(time, template.__name__, name);
+                return result;
+            })
+                // @TODO: need to handle errors with log
+
+                .catch(e => {
+                    getStats(time, template.__name__, name, true);
+                    throw e;
+                });
         }
 
         // Add persistors to foreign key references
-        if (template.defineProperties && typeof(template._injectProperties) == 'function')
+        if (template.defineProperties && typeof (template._injectProperties) == 'function')
             template._injectProperties();
 
     };
@@ -699,7 +990,7 @@ module.exports = function (PersistObjectTemplate, baseClassForPersist) {
      * @returns {object} returns transaction object
      */
     PersistObjectTemplate.begin = function (notDefault) {
-        var txn = {id: new Date().getTime(), dirtyObjects: {}, savedObjects: {}, touchObjects: {}, deletedObjects: {}};
+        var txn = { id: new Date().getTime(), dirtyObjects: {}, savedObjects: {}, touchObjects: {}, deletedObjects: {} };
         if (!notDefault) {
             this.currentTransaction = txn;
         }
@@ -707,10 +998,10 @@ module.exports = function (PersistObjectTemplate, baseClassForPersist) {
     };
 
 
-    PersistObjectTemplate.end = function(persistorTransaction, logger) {
+    PersistObjectTemplate.end = function (persistorTransaction, logger) {
         persistorTransaction = persistorTransaction || this.currentTransaction;
         logger = logger || PersistObjectTemplate.logger;
-        return PersistObjectTemplate.commit({transaction: persistorTransaction, logger: logger});
+        return PersistObjectTemplate.commit({ transaction: persistorTransaction, logger: logger });
     };
     /**
      * Set the object dirty along with all descendant objects in the logical "document"
@@ -745,7 +1036,7 @@ module.exports = function (PersistObjectTemplate, baseClassForPersist) {
             // Potentially cascade to set other related objects as dirty
             topObject = PersistObjectTemplate.getTopObject(obj);
             if (!topObject) {
-                (logger || this.logger).error({component: 'persistor', module: 'api', activity: 'setDirty'}, 'Warning: setDirty called for ' + obj.__id__ + ' which is an orphan');
+                (logger || this.logger).error({ component: 'persistor', module: 'api', activity: 'setDirty' }, 'Warning: setDirty called for ' + obj.__id__ + ' which is an orphan');
             }
             if (topObject && topObject.__template__.__schema__.cascadeSave) {
                 PersistObjectTemplate.enumerateDocumentObjects(PersistObjectTemplate.getTopObject(obj), function (obj) {
@@ -791,6 +1082,7 @@ module.exports = function (PersistObjectTemplate, baseClassForPersist) {
     };
 
     PersistObjectTemplate.saveAll = async function (txn, logger) {
+        var time = getTime();
         var promises = [];
         var somethingSaved = false;
         var dirtyObjects = txn ? txn.dirtyObjects : this.dirtyObjects;
@@ -817,7 +1109,16 @@ module.exports = function (PersistObjectTemplate, baseClassForPersist) {
                 } else {
                     return true;
                 }
-            }.bind(this));
+            }.bind(this))
+            .then(result => {
+                getStats(time, 'PersistObjectTemplate', 'saveAll');
+                return result;
+            })
+            // @TODO: need to handle errors with log
+            .catch(e => {
+                getStats(time, 'PersistObjectTemplate', 'saveAll', true);
+                throw e;
+            });
     }
 
     /**
@@ -830,7 +1131,7 @@ module.exports = function (PersistObjectTemplate, baseClassForPersist) {
         type = type || PersistObjectTemplate.DB_Mongo;
         alias = alias || '__default__';
         this._db = this._db || {};
-        this._db[alias] = {connection: db, type: type}
+        this._db[alias] = { connection: db, type: type }
     };
 
     /**
@@ -848,7 +1149,7 @@ module.exports = function (PersistObjectTemplate, baseClassForPersist) {
             PersistObjectTemplate.getPOJOFromMongoQuery(template, query, options, logger) :
             PersistObjectTemplate.getPOJOsFromKnexQuery(template, [], query, options, undefined, logger).then(function (pojos) {
                 pojos.forEach(function (pojo) {
-                    _.map(pojo, function(_val, prop) {
+                    _.map(pojo, function (_val, prop) {
                         if (prop.match(RegExp('^' + prefix + '___'))) {
                             pojo[prop.replace(RegExp('^' + prefix + '___'), '')] = pojo[prop];
                             delete pojo[prop];
@@ -860,29 +1161,44 @@ module.exports = function (PersistObjectTemplate, baseClassForPersist) {
     }
 
     PersistObjectTemplate.beginTransaction = function () {
-        var txn = {id: new Date().getTime(), dirtyObjects: {},
-            savedObjects: {}, touchObjects: {}, deletedObjects: {}, deleteQueries: {}};
+        var txn = {
+            id: new Date().getTime(), dirtyObjects: {},
+            savedObjects: {}, touchObjects: {}, deletedObjects: {}, deleteQueries: {}
+        };
         return txn;
     };
 
     PersistObjectTemplate.beginDefaultTransaction = function () {
-        this.__defaultTransaction__ = {id: new Date().getTime(), dirtyObjects: {}, savedObjects: {}, touchObjects: {}, deletedObjects: {}};
+        this.__defaultTransaction__ = { id: new Date().getTime(), dirtyObjects: {}, savedObjects: {}, touchObjects: {}, deletedObjects: {} };
         return this.__defaultTransaction__;
     };
 
     PersistObjectTemplate.commit = async function commit(options) {
+        var time = getTime();
         PersistObjectTemplate._validateParams(options, 'commitSchema');
 
         options = options || {};
         var logger = options.logger || PersistObjectTemplate.logger;
 
-                //var dbType = PersistObjectTemplate.getDB(PersistObjectTemplate.getDBAlias(template.__collection__)).type;
-                //var prefix = PersistObjectTemplate.dealias(template.__collection__);
+        //var dbType = PersistObjectTemplate.getDB(PersistObjectTemplate.getDBAlias(template.__collection__)).type;
+        //var prefix = PersistObjectTemplate.dealias(template.__collection__);
         var persistorTransaction = options.transaction || this.__defaultTransaction__;
 
+        var promise;
         if (PersistObjectTemplate.DB_Knex) {
-            return PersistObjectTemplate._commitKnex(persistorTransaction, logger, options.notifyChanges);
+            promise = PersistObjectTemplate._commitKnex(persistorTransaction, logger, options.notifyChanges);
         }
+
+        const name = 'commit';
+        return promise.then(result => {
+            getStats(time, 'PersistObjectTemplate', name);
+            return result;
+        })
+            // @TODO: need to handle errors with log
+            .catch(e => {
+                getStats(time, 'PersistObjectTemplate', name, true);
+                throw e;
+            });
     };
 
     /**
@@ -894,7 +1210,7 @@ module.exports = function (PersistObjectTemplate, baseClassForPersist) {
     PersistObjectTemplate.connect = function (config, schema) {
         var knex = require('knex');
         var connection = knex(config);
-        this.setDB(connection, this.DB_Knex,  config.client);
+        this.setDB(connection, this.DB_Knex, config.client);
         this.setSchema(schema);
         this.performInjections(); // Normally done by getTemplates
         return connection;
@@ -928,12 +1244,12 @@ module.exports = function (PersistObjectTemplate, baseClassForPersist) {
     PersistObjectTemplate.onAllTables = function (action, concurrency) {
         var templates = [];
         _.each(this.__dictionary__, drop);
-        function drop (template) {
+        function drop(template) {
             if (template.__schema__ && (!template.__schema__.documentOf || !template.__schema__.documentOf.match(/not persistent/i))) {
                 templates.push(template);
             }
         }
-        return Promise.map(templates, action, {concurrency: concurrency || 1});
+        return Promise.map(templates, action, { concurrency: concurrency || 1 });
     }
 
 };